requirements.txt
streamlit
folium
streamlit-folium
requests
python-docx
Pillow
selenium
chromedriver-autoinstaller